renamed scraper to query
