+++ splatnet3_scraper/__init__.py
@@ -1,10 +1,6 @@
 import pathlib
 import re
 
-__version__ = "0.4.5"
+__version__ = "0.5.0"
 
-from splatnet3_scraper.scraper import (
-    JSONParser,
-    QueryResponse,
-    SplatNet3_Scraper,
-)
+from splatnet3_scraper.query import JSONParser, QueryResponse, SplatNet3_Scraper
