+++ splatnet3_scraper/query/config.py
