+++ splatnet3_scraper/query/query_map.py
