--- splatnet3_scraper/scraper/__init__.py
@@ -1,3 +1,0 @@
-from splatnet3_scraper.scraper.json_parser import JSONParser
-from splatnet3_scraper.scraper.main import SplatNet3_Scraper
-from splatnet3_scraper.scraper.responses import QueryResponse
